drawer successfully renders grantlist
