+++ grantly/src/components/grants/GrantList.js
@@ -6,60 +6,10 @@ import { connect } from "react-redux";
 import Grant from "./Grant";
 import Loader from "react-loader-spinner";
 import { fetchApi } from "../../actions";
-import Drawer from "@material-ui/core/Drawer";
-import { makeStyles, useTheme } from "@material-ui/core/styles";
-import IconButton from "@material-ui/core/IconButton";
-import MenuIcon from "@material-ui/icons/Menu";
-import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
-import ChevronRightIcon from "@material-ui/icons/ChevronRight";
-const drawerWidth = 240;
 
 // Styles
-const useStyles = makeStyles(theme => ({
-  root: {
-    display: "flex"
-  },
-  drawer: {
-    [theme.breakpoints.up("sm")]: {
-      width: drawerWidth,
-      flexShrink: 0
-    }
-  },
-  appBar: {
-    marginLeft: drawerWidth,
-    [theme.breakpoints.up("sm")]: {
-      width: `calc(100% - ${drawerWidth}px)`
-    }
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-    [theme.breakpoints.up("sm")]: {
-      display: "none"
-    }
-  },
-  toolbar: theme.mixins.toolbar,
-  drawerPaper: {
-    width: drawerWidth
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3)
-  }
-}));
 
 const GrantList = props => {
-  const classes = useStyles();
-  const theme = useTheme();
-  const [open, setOpen] = React.useState(true);
-
-  function handleDrawerOpen() {
-    setOpen(true);
-  }
-
-  function handleDrawerClose() {
-    setOpen(false);
-  }
-
   console.log("GrantList props", props);
   useEffect(() => {
     if (props.data.length === 0) {
@@ -72,17 +22,8 @@ const GrantList = props => {
     return <Loader type="Triangle" color="#00BFFF" height="100" width="100" />;
   }
   // console.log(user);
-
   return (
-    <Drawer
-      className={classes.drawer}
-      variant="persistent"
-      anchor="left"
-      open={open}
-      classes={{
-        paper: classes.drawerPaper
-      }}
-    >
+    <div>
       {props.data.length > 0 ? (
         props.data.map(grant => {
           return <Grant grant={grant} key={grant.id} />;
@@ -90,8 +31,7 @@ const GrantList = props => {
       ) : (
         <div> Grants incoming! </div>
       )}
-      <div className={classes.drawerHeader}></div>
-    </Drawer>
+    </div>
   );
 };
